Collision Detection Bug Fix
made a simple mistake, used "contains" instead of "intersects" for both environment and object. This fixes instability on object-object as well as left wall collision.

Soon I will consolidate the methods and clean up naming conventions, as well as do a spicier friction handling, as the current friction feature is only for flash. I want to solidify friction in the game to give it better feel.